requirements.txt
click>=8.1.7
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.70.0,<2.0.0
langchain-core==0.3.74
langchain-openai==0.2.14
langsmith>=0.3.18,<0.4.0
tiktoken>=0.5.0,<1.0.0
numpy==1.26.4
pandas==2.2.2
pypdf>=4.2.0,<6.0.0
google-api-core==2.11.0
google-api-python-client==2.131.0
google-auth==2.22.0
google-cloud-aiplatform==1.53.0
google-generativeai==0.5.4
protobuf==4.24.4   
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-proto==1.18.0  
opentelemetry-exporter-otlp-proto-common==1.18.0  
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
